Add details page for msgs and error page fallback
A default page has been made for the route /details
which will be the most recent message. Otherwise
when the open button is clicked next to messages
on the wall it will open a details page using
a query to pass the data to the details view.

In the event that a message doesn't exist an error
page appears to let the user know.